refactor(computer-use): give each platform its own leaf directory
The leaves sat side by side in one directory: window.rs, capture.rs, uia.rs and
input.rs were Windows-only, platform_macos.rs held the macOS equivalent of all
four, and nothing in the names said which was which. A reader had to know the
convention to tell them apart, and the macOS file had grown to 1,229 lines
covering window discovery, capture, accessibility and input at once.

Each platform now owns a directory, and macOS is split the way Windows already
was -- window, capture, accessibility_tree, input -- so the same concern lives
at the same place on both sides. Helpers used by more than one macOS leaf (the
CoreGraphics dictionary readers, window geometry, the private AX window lookup)
sit in that platform's mod.rs, so leaves depend on their platform module rather
than on each other.

Directories are named platform_windows and platform_macos rather than windows
and macos: a module named windows would shadow the windows crate the leaves
import from.

Function bodies are moved verbatim -- the macOS split was carved by line range,
and the before/after inventory is the same forty-five functions. Leaf items
widen from pub(super) to pub(crate) because their platform module now re-exports
them one level up. Windows compiles clean with no warnings.
