+++ console/src-tauri/src/computer_use_server/platform_windows/mod.rs
@@ -1,0 +1,20 @@
+//! Windows Computer Use leaves.
+//!
+//! The RPC server, framing, transport, session/turn lifecycle, and app approval
+//! are shared with macOS (see the parent `mod.rs`). Only the OS-touching leaves
+//! below are platform specific: window discovery over EnumWindows, capture
+//! through Windows Graphics Capture, accessibility over UI Automation, and
+//! input over SendInput.
+
+mod capture;
+mod input;
+mod uia;
+mod window;
+
+pub(super) use capture::observe_window;
+pub(super) use input::{
+    click, desktop_locked, drag, press_key, scroll, set_focus, set_intervention_bypass_once,
+    type_text,
+};
+pub(super) use uia::{invoke_element, set_value};
+pub(super) use window::{close_window, is_forbidden, list_apps, list_windows, resolve_window};
